+++ R/qctoolsPackage.R
@@ -1,0 +1,11 @@
+#' @keywords internal
+"_PACKAGE"
+
+globalVariables(
+  c(
+    "is_header",
+    "sha",
+    "commit",
+    "author"
+  )
+)