+++ NEWS.md
@@ -1,3 +1,9 @@
+# qctools development
+
+## New features and changes
+
+- `repoHistory` added to return all commits made to every file. (#32)
+
 # qctools 0.2.1
 
 ## Bug fixes